--- a/daml-lf/interpreter/src/main/scala/com/digitalasset/daml/lf/speedy/Anf.scala
+++ b/daml-lf/interpreter/src/main/scala/com/digitalasset/daml/lf/speedy/Anf.scala
@@ -99,11 +99,15 @@ object Anf {
   type K[T] = ((DepthA, T) => Res)
 
   /** During conversion we need to deal with bindings which are made/found at a given
-    absolute stack depth. These are represented using `AbsBinding`.
+    absolute stack depth. These are represented using `AbsBinding`. An absolute stack
+    depth is the offset from the bottom of the stack, which is empty when a function is
+    entered. We call it absolute because an offset doesn't change as new bindings are
+    pushed onto the stack.
 
-    Note the contrast with the expression form `ELocS` which indicates a relative offset
-    from the top of the stack. This relative-position is used in both the original
-    expression which we traverse AND the new ANF expression we are constructing.
+    Note the contrast with the expression form `ELocS` which contains a relative offset
+    from the end of the stack. This relative-position is used in both the original
+    expression which we traverse AND the new ANF expression we are constructing. The
+    relative-offset to a binding varies as new bindings are pushed on the stack.
     */
   case class AbsBinding(abs: DepthA)