ANF transformation in Speedy.

The idea behind this PR is to transform speedy expressions into a simpler form where all non-atomic sub-expressions are made explicit by the introduction of let-forms. In particular, for the function-application form. These simpler forms allow the execution engine to take advantage of the atomic assumption, and often removes many additional execution steps. In particular the pushing of continuations to allow execution to continue after a compound expression has been reduced to a value.


The performance is increase by around x1.15. Here are the raw figures for the CollectAuthority benchmark, run 6x, interleaving between the master branch and this branch.

```
anf:     64.538 ±(99.9%) 0.442 ms/op [Average]
master:  75.250 ±(99.9%) 0.444 ms/op [Average]
anf:     64.160 ±(99.9%) 0.674 ms/op [Average]
master:  73.407 ±(99.9%) 0.850 ms/op [Average]
anf:     63.549 ±(99.9%) 0.753 ms/op [Average]
master:  73.293 ±(99.9%) 0.500 ms/op [Average]
```

This PR relates to issue #6198 

changelog_begin
changelog_end