@nickchapman-da 2020-06-25
> Why has this PR gotten so much bigger since Monday when it was almost ready?

We made the change to distinguish `SExpr` and `AExpr`