@nickchapman-da 2020-06-30
Address new comments from @remyhaemmerle-da