@nickchapman-da 2020-06-30
Addressed comments/questions from @hurryabit  which github had hidden from me.